python/src/requirements.txt
attrs ~=19.1
inflection ~=0.3
jsonschema ~=3.0
six ~=1.0
typing ~= 3.7
